requirements.txt
deepgram-sdk==3.2.1
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
# only used for python Microphone
PyAudio==0.2.14
